Rename "resolve" to "update"
